+++ CHANGELOG.md
@@ -7,7 +7,9 @@ from version 1.20.0 onwards.
 
 ## [Unreleased]
 
-- n/a
+### Added
+- map/flat_map now accept an `error_fn` to transform the result of an
+  unsuccessful future ([#153](https://github.com/rohanpm/more-executors/issues/153)).
 
 ## [2.1.2] - 2019-06-26